requirements.txt
appnope==0.1.0
argon2-cffi==20.1.0
astroid==2.4.2
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
cffi==1.14.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.0
numpy==1.19.1
opt-einsum==3.3.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pi==0.1.2
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pyro-api==0.1.2
pyro-ppl==1.4.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pyzmq==19.0.2
qtconsole==4.7.5
QtPy==1.9.0
scipy==1.1.0
Send2Trash==1.5.0
six==1.15.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
